+++ CONTRIBUTING.md
@@ -12,7 +12,7 @@
 docker build -t rate-api . && docker run --name rate-api -t -i -e DATABASE_URL=postgresql://user:pwd@dburl/exchange rate-api 
 ```
 
-# Refactor and Improvement
+# todo Refactor and Improvement
 
 - [ ] unit test
 - [ ] integration test
